wip: raw sign and submit e2e - works on preprod

+++ nrf52-demo/demo-client/src/types.rs
@@ -1,5 +1,6 @@
 
+#[derive(Debug)]
 pub struct DeviceData {
-  pub sensor_readings: i8,
+  pub sensor_readings: i32,
   pub signed_readings: Vec<u8>,
 }
